[OLINGO-1191]Code improvements

+++ lib/client-core/src/main/java/org/apache/olingo/client/core/communication/request/AsyncRequestWrapperImpl.java
@@ -203,7 +203,7 @@ public class AsyncRequestWrapperImpl<R extends ODataResponse> extends AbstractRe
 
           try {
             // wait for retry-after
-            Thread.sleep(retryAfter * 1000);
+            Thread.sleep((long)retryAfter * 1000);
           } catch (InterruptedException ignore) {
             // ignore
           }
